Small ConsPrefShock changes + kernel regression
Twiddled the ConsPrefShock solver and simulator a bit, so that
simulation now works.  Maybe it always worked, and there wasn't a target
wealth when I thought there was.

Also added (but have not tested) a basic kernel regression method to
HARKutilities; need data on other computer, thus the commit.  Should add
kernels other than Epanechnikov and generalize kernelRegression to allow
this.