Changed simulation variables names
Forgot to change the variable names in simulating and
shock-history-generating methods and functions for previous commit, now
fixed.  SolvingMicroDSOPs and cstwMPC (the files that rely on
simulation) both run.

A bunch of variables in cstwMPC still need to be renamed, but this can
wait until the simulation is revamped in the near future (the "effective
interest rate" as permanent shock needs to come out).