Many ConsumptionSavingModel (and core) improvements
Assorted changes and fixes as previously planned:

- HARKcore now has a Solution class that individual model solutions can
inherit from if they want.  Its only feature is a "default distance"
method (which can be overwritten, obv).  To use the method, the Solution
subclass merely needs to have an attribute called convergence_criteria,
which lists the attributes that must be sufficiently close for an
infinite process to be considered "converged".

- The distance method of ConsumerSolution has been deleted and replaced
with self.convergence_criteria = ['cFunc'], which does the same thing.
You can see how human wealth matters in infinite horizon models by
adding 'hRto' to convergence_criteria.

- The perfect foresight solver now takes LivFac as an input, consistent
with all other models in its module.

- The Markov solver generalization is now working correctly in seemingly
all cases.  It can deal with the interest rate, permanent income growth
factor, and income distribution varying across states; now correctly
handles MPC limits and borrowing constraints in oddball cases.

- MPC limits and human wealth are always computed now, rather than only
when CubicBool or vFuncBool are True.  However, they are not actually
used unless one of those is true.

- The "exogenous grid" solver has been deleted, per agreement (along
with the "hybrid type" example).  It can be re-added or rewritten later
(in OOP format), but was hopelessly outdated at this point.

- Value function construction now seems to work correctly in all models
in ConsumptionSavingModel: perfect foresight, risky income, kinked R,
and Markov (even in oddball cases).  It's a bit slow right now.

- The distance method of CubicInterp now depends on dydx values at nodes
(along with x and y values), and does not depend on extrapolation
choice.