Gen eq test model: FashionVictim
The general gen eq solver now has a "baby" test model to work with,
located in the /FashionVictim folder.  It is a model of dynamic binary
fashion choice, inspired by the paper, "Why Do All Hipsters Look The
Same?".  The Market.solve() method can find a dynamic equilibrium of
this model for some parameters, but convergence of the dynamic rule
doesn't always happen.

The old approxMeanOneLognormal (commented out) has now been deleted, as
it's just wasted space.

The AgentType class now has seed and RNG as standard attributes, along
with the resetRNG() method.  Going to incorporate this into existing
applications on next commit.