Three interpolation improvements
1) LinearInterp now has the option for "decay extrapolation" above the
range of x-points, just like CubicInterp.  Consumption functions now
respect perfect foresight solutions even when linear interpolation is
used.

2) CubicInterp was partially rewritten, resulting both in simpler code
and a significant speedup-- maybe 3x or more?

3) LowerEnvelope has been rewritten entirely so that it's no longer a
lie.  This was formerly ConstrainedComposite but was renamed during
Namingway, with no change in functionality.  Now it does what it says on
the tin: you pass it any number of 1D functions, it makes a function
that always returns the lowest among them.