Made new Sphinx docs after colectomy
The Sphinx docs are now much more complete, but there are still some
oddities.  The __call__ method for classes is in the summary for that
class, but its full description is not provided.  And the __init__
method is excluded from both the summary and the full documentation;
this is pretty important for a lot of our classes.