Edited top-of-file docstrings
The topline docstring for every module has been updated.  As a style
choice, the first sentence of each docstring lacks a subject, which is
assumed to be "this module".  This is because the first sentence will
appear in the summary/index in Sphinx documentation, and this
circumvents ugly repetition and is more succinct.  Suggest we do this
for functions and methods as well.

ConsPrefShockModel docstring might need to be edited if I manage to make
some more changes to it.