Explicit PermInc solver working
The idiosyncratic shocks with explicit permanent income solver seems to
be working correctly.  It is about 8-9x faster when the lower bound of m
is zero (artificially or naturally), as this enables it to use
LinearInterpOnInterp1D rather than Curvilinear2DInterp.

Added LowerEnvelope2D and a robustness fix for Curvilinear2DInterp
(which can / should still be improved).  findCoords() would fail for
trapezoid sectors because math; there's an alternate algebraic solution
for (alpha,beta) if the sector is a trapezoid, now implemented.