Simulation for persistent shocks works
Small updates to get persistent shock model simulation working.  Note
that the "naturally constrained consumption function" is a
HARKinterpolator2D with an empty distance_criteria ... This means that
convergence in an infinite horizon model *will not* depend on this
portion of the consumption function, only the "artificially constrained"
portion and the unconstrained portion.

To do: update makePermIncGrid to choose a better grid when there are
persistent shocks.  Requires math.