Fixed med terminal, added med simulation
The terminal solution in the medical shocks model has been adjusted
correctly.  Takes nearly as long to solve terminal period solution as
non-terminal period, which is worrisome because a version of the former
must be done each period in the presence of a binding borrowing
constraint.

Added simulation routines for medical shock model; reinforces that we
need another simulation overhaul.  Test runs show that agents
occasionally (like, 1 in a million or less) violate their budget
constraint when they get a very bad medical shock.  This is a result of
badly capturing the *convexity* of the consumption function at the very
bottom.  Might need denser grid at bottom... or switch to cubic splines.

Programmed some of a bicubic interpolation method, untested.  Wrote
constructor and interpolation part, but there are *eight* extrapolation
routines to write.  Ugh.