MedShocks borrowing / constrained solution works
Until now, ConsMedModel forced a natural borrowing constraint of zero by
inserting a 0 income shock with tiny probability.  This got around the
difficulties with borrowing and constraints in a two good model.  Now
the model works like the others, with an arbitrary borrowing constraint
allowed.

Also realized that the solver is much slower than I thought, because I
had the wrong number of gridpoints selected.