cFromx function has cubic spline capability
The "consumption from expenditure" function now uses cubic splines in
the xLvl dimension when CubicBool=True.  This did not fix the wiggliness
issues with a small number of gridpoints and cubic spline interpolation.
On the upside, cubic interpolation doesn't generate as great a time
increase (as a factor) than in other models.