Merged PR #86
Modified HARKparallel.py so that it could be imported as a module even if joblib and dill are not installed.  Functions that do not use joblib and dill can be called.  Functions that do use joblib and dill, if called, will raise an error message indicating that joblib and dill are not installed.