Check Rboro is greater than Rsave for kinky consumers
The solution to the kinky consumer behaves weird when Rboro=Rsave. For
example, in ‘ConsumerParameters.py’ set both rates equal to 1.02, then
run ‘ConsIndShockModel.py’ and look at the (kinky) consumption function
plotted.
The fix tells the user that when Rboro=Rsave, it is better to use the
model without kinks. The fix is introduced in the function
‘solveConsKinkedR’, immediately after the code checks that Rboro>=Rsave.