Fix issue #100
Pavel Solis pointed out in #100 that the KinkedR solver breaks when
Rboro == Rsave.  This is due to duplicative interpolation points, which
breaks LinearInterp and causes nonsense output.

The fix is to put a conditional around the offending lines, which makes
the solver act just like ConsIndShockSolver.  I added some comments to
this effect; also moved the Rboro >= Rsave assert to be with the
CubicBool assert for aesthetics.