Fixed most warnings, tested simulation
All but two of the warnings that pop up when solving BabyLaborModel have
been resolved.  One of the remaining ones comes up because of
duplicative points in a LinearInterpDiscont's x_list, which might be due
to either a) a bug in FellaInterp.addNewPoints or b) extraordinary bad
luck.

Added a few lines to __main__ block to demo simulation and plot labor
supply rate by age.  It works.