Added Fagereng, Holm, & Natvik demo
MNW & CDC discussed a paper by Andreas Fagereng et al at CESifo 2017 in
Venice, including a quick and dirty structural estimation to see if some
of that paper's results could be explained by a simple
consumption-saving model.  That demo is now in /ConsumptionSaving/Demos/

Also put one line fix in HARKsimulation because of an update to numpy.
This replicates the fix in the ConsAggShockFix PR.