Added ConsIndShockSolverRobust
Wrote a solver function (and accompanying solver class) for
ConsIndShockModel that is robust to non-concave future value, and adds
the option of a consumption floor (normalized to permanent income).
Doesn't handle cubic interpolation, and value function is constructed
automatically (necessary with non-concave value).  Untested, will make
and test example for next commit.