Added RobustConsumerType, solver sorta works
solveConsIndShockRobust kinda sorta works with the new ConsumerType
subclass.  It works for low levels of cFloor, but raising cFloor to the
point where it's actually *relevant* (i.e. can be experienced by a
consumer who has a valid a_t and gets some shock from the distribution)
causes the solver to fail.  This seems to occur because of large
discontinuities in EndOfPrdvP from discrete shocks crossing the boundary
from v' = 0 (on cFloor) to v' >> 0 (above cFloor).  This causes the
candidate set to jag up and down repeatedly, which can result in an
(economically / theoretically) impossible value function being
constructed by FellaInterp.  This occurs when moving_right flips from
False to True while the candidate set dominates, leading to a value
discontinuity.  This might require a reworking of
FellaInterp.addNewPoints.

Moving on to ChildLaborModel for now, which shouldn't have these
problems because the changes in v' are much smaller.