Reworked ConsIndShock to allow extension to cFloor
Some methods in ConsIndShockSolver had to be rearranged in order to
accommodate the cFloor solver (formerly Robust).  All models seem to
run.  Had to temporarily introduce a small mistake: TranShkDstn is in
time_inv, but should be in time_vary.  The functionality to make this
happen automatically is in the branch AggShockMarkov.  Must merge dev
branches to fix.