Poly-state "jumper" works
ConsAggMarkov can now handle models with many Markov states through the
use of a "state jumper" when constructing the sequence of states.  Until
each state has been reached a certain number of times, act_T more
periods are added to the sequence, and we "jump" to an underrepresented
state.  Might have a bug, investigating.