StickyE working with SOEmarkov
The SOE-Markov model is working in StickyE.  Had to give newborns 1.0
for TranShkNow, else change in log consumption was crazy.  Added
IdentityFunction to HARKinterpolation, which is a copy-paste from a
private MNW branch.  Will clean up the main file to make it a bit more
elegant.  Everything should be set for HA-DSGE versions, will do
tomorrow.