Checking the conditions for nondegenerate solutions  (#115)
* Check Impatience Conditions

A method that checks 6 conditions, which may be sufficient for a
nondegenerate solution.

* Corrections

All parameters are of the instance’s type. A check for infinite horizon
models with a 1 period cycle is also added.

* Adding [0] for a list variable

PermGroFac[0]

* Corrections

RIC, AIC, and FHWC are evaluated by PerfForesightConsumerType while
GIC, WRIC, and FVAC are evaluated by IndShockConsumerType. The code is
restructured to evaluate one condition at a time in a block. The method
is renamed to checkConditions. Extra indentation is avoided.

* Corrections

T_total is replaced with T_cycle. Calculation of exp_psi_inv and
exp_psi_to_one_minus_rho is omitted from PerfForesightConsumerType.
Returns feedback is changed with none in the docstrings. Default
verbosity is false.