Improve portability, clean up tables
Stata-calling code should now be fully portable, as long as user sets the location of the executable correctly.  Made various small changes to tables to fix typos / formatting.

Re-ran the models with 20k periods and 20k agents, but might need to redo again.  Seem to be using wrong version of yLvl in the simulation-- includes capital and labor income, but paper only has y=labor income.