Added cost-of-stickiness exercise
When calc_micro_stats is True, the simulation processing function now calculates the cost of stickiness as the willingness to pay (in terms of fractional reduction in permanent income) at birth to *not* be sticky.  These are reported in the equilibrium table, aggregated across all states.  It is possible to recover state-by-state WTP from the files generated.  I think we have to choose a different method for the DSGE model, as we're currently picking up individual WTP for *everyone* to not be sticky.

Still need to edit the micro regressions so that they don't cause memory problems when running with many agents and periods.  Still commented out.