Made and ran alt uCost, has serious bug
Alternate cost of stickiness exercise for HA-DSGE has been set up and run, but I made a (big) mistake somewhere.  The cost of stickiness is now on the order of 2.4% of permanent income, and varies wildly by state.  High growth states have ridiculously negative cost of stickiness, which is implausible.  I failed to save the data file, so I need to re-run it to see what's happening.