Simplified RA simulation methods
The RA class had a weird getStates method that did a bunch of wasted work, now cleaned up.  And the getShocks method was using an old version of the model in which the rep agent only sees a Pi portion of the transitory shock, which is not what's in the paper.  Now fixed.

Also added no measurement error tables to AllTables.pdf, so they're easy to find.