Added representative agent models
Even though it's a straight up contradiction, I've added two simple representative agent models in a new module.  These were written for the cAndCwithStickyE paper, which will soon be added to main HARK as an additional paper example (the version in the repo now is old and broken and was only meant to be added temporarily).