Delete ConsPersistentShockModel and ConsMedModelOLD
Removes old modules that have been replaced on this branch.  As of *just before* this commit, these legacy models worked properly (as long as PrstIncCorr was re-named to PermIncCorr in the parameter dictionary).

Also removes a few dictionary lines from ConsumerParameters.py that were used by the old versions but not the new versions.