Fixed imports in model files
All of the consumption-saving model files used a style of import that didn't work when the file was run directly (rather than called as a module), even though they have a __main__ block (and main() function).  This has now been fixed.

Also mostly removed extraneous file RepAgentModel.py, which seems to be an old name of ConsRepAgentModel.py.  This file now simply imports all of ConsRepAgentModel and warns the user to use that instead.