Changed hardcoded updateAFunc parameters into proper parameters
Four parameters that govern how CobbDouglasEconomy.updateAFunc works
were defined locally, within the method, but are now attributes to be
assigned at init (or at least before the user tries to solve):

- update_weight --> DampingFac, now defined complementarily
- verbose --> verbose
- discard_periods --> T_discard
- max_loops --> max_loops

To prevent this from being a breaking change, init method writes old
hardcoded values if omitted from passed inputs.  This will be improved
with a warning later.

Untested, as it turns out Anaconda3 is incorrectly installed on this
computer.