Add liquidity constraints to perfect foresight model
PerfForesightConsumerType can now handle an artificial borrowing constraint, like in all our other models.  Currently doesn't calculate value function if BoroCnstArt is not None, as this requires additional work to compute.

When this is merged, it will be a breaking change for all existing instances of PerfForesightConsumerType, as they will now require BoroCnstArt and aXtraCount to be attributes.