Fix one line in distance metric
The "universal distance metric" had a bug in one line that had never come up.  If the two objects being compared have a shape, but the shapes are not the same, then it tried to take the supnorm between them; this fails because arithmetic is not defined on tuples.  Fix simply converts the shape tuples into arrays.