Add default values for BoroCnstArt and MaxKinks
If BoroCnstArt is not specified in PerfForesight model, preSolve assumes user meant no artificial borrowing constraint (as the model has always been until now).  If BoroCnstArt is specified (not None) *and* it's an infinite horizon model *and* MaxKinks is not specified, then an AttributeError is raised to notify the user.