Implemented faster expectations in new portfolio model
If IncomeDstn and RiskyDstn are independent, then expectations over shocks can be computed sequentially.  This speeds up the code by roughly a factor of RiskyCount.  The solver can also correctly handle non-independent distributions, but as is the update method makes them independent (and sets the appropriate flag).

I've discovered that the solver becomes unstable when UpdatePrb is low, like 0.1.  It works for 0.15.  Not sure what's going on.