Changed test target values
Just remembered that the problem is that PKM's ShareFunc was defined on aNrm, but mine is defined on mNrm.  That's what was causing the discrepancy.  Had to just change the target values to get it to work, because there's no way to reproduce Patrick's target numbers in the new version.