Swap portfolio models (#594)
* Remove old ConsPortfolioModel

Swap ALT for previous ConsPortfolioModel.py.  Examples run as intended.  Remove files that depended on old ConsPortfolioModel.

* Fix function name in CPM test

Solution attribute names change from old to new version.  Updated references in test.

* Make test ensure CPM type is infinite horizon

* Changed test target values

Just remembered that the problem is that PKM's ShareFunc was defined on aNrm, but mine is defined on mNrm.  That's what was causing the discrepancy.  Had to just change the target values to get it to work, because there's no way to reproduce Patrick's target numbers in the new version.