Fix DiscreteDistribution.drawDiscrete
The new version of drawDiscrete did not work properly when the distribution was multivariate.  That's now fixed, returning a JxN array of draws (where J is the dimension of the distribution).

It also seemed to have a typo/error, referring to Xdim, which does not exist.

This commit does not make drawDiscrete work when the input X is an integer, but it didn't work before either.