Revise test_ConsAggShockModel
Tests for models in ConsAggShockModel.py have been revised and expanded.  There are now tests for "micro" and "macro" versions of both the regular and Markov versions.

Options have been set so that it "fake solves", with a small number of agents, short history, and simply quitting on solving after 3 loops.