Add missing parentheses in ConsPFmodel
A set of parentheses was missing in line 514, which handles the special case of the borrowing constraint binding at *all* of the cFunc kinks.  This screwed up the solution to simple problems (literally: the very first consumption-saving problem a student would ever do!) but had no effect on infinite horizon solutions.