interpolation.py: Before extrapolating, test whether already converged
Our 1D interpolation routine allows the user to specify a line toward which the function is assumed to be decaying at an exponential rate beyond the uppermost explicit gridpoint.   A degenerate case that throws a difficult-to-decipher bug occurs when the last segment of the linear interpolation is ALREADY exactly (to numerical precision) on the limiting line.  This tests for that possibility, and if true, just reverts to linear extrapolation.