dcegm: Choices -> choices (#979)
* Fix plotFuncs -> plot_funcs etc

* Handle degenerate case: ConsIndShockModel solved with zero uncertainty

* Fix tests to correspond to fixed code

* interpolation.py: Before extrapolating, test whether already converged

Our 1D interpolation routine allows the user to specify a line toward which the function is assumed to be decaying at an exponential rate beyond the uppermost explicit gridpoint.   A degenerate case that throws a difficult-to-decipher bug occurs when the last segment of the linear interpolation is ALREADY exactly (to numerical precision) on the limiting line.  This tests for that possibility, and if true, just reverts to linear extrapolation.

* Fix capitalization error in dcegm.py

* usePointsForInterpolation -> use_points_for_interpolation