Add logging, autorun calc_stable_points
The conditions_report is now sent to _log.info if not quiet. ConsIndShock.post_solve will now run calc_stable_points() if appropriate, so that the target and balanced-growth mNrm levels are added to the solution.

At the default parameters (infinite horizon), the main branch takes 0.328 seconds to solve on my computer. After moving the calc_stable_points code outside of the solver loop and only running it at the end (if appropriate), it takes 0.100 seconds to solve. Over two-thirds of solution time was being *wasted* on calc_stable_points, whose work is not used during the solution and *could not* plausibly be used.