Patience factors (etc) now in self.auxiliary
A new dictionary has been added to PerfForesightConsumerType and IndShockConsumerType, called auxiliary. It contains all the various patience factors and other semi-useful values that are constructed for check_conditions, but aren't needed to solve (nor simulate) the model. This dictionary can be renamed with a simple replace-all on self.auxiliary.