Add simplified PF solver
Adds a non-OO solve_one_period function for the ConsPerfectForesight model. Will do for the other solvers as well.

The value function in the PF model isn't right when there's an artificial borrowing constraint (in current HARK). This commit makes it less wrong, but I don't think it's *right* yet.