Add simple solver for ConsMedShockModel
The "medical shock" model now has a simple solver that replicates behavior of the existing solver. The current OO version's CubicBool capability is broken; it works in the new simple solver, but I haven't fixed it in the OO code.

It looks like the tests for this model just make sure that it can run and simulate, but I've verified that the solution reproduces exactly on my own. As with other models, this solver runs faster than before, but not by as much: 0.5-2.5x faster.

There are improvements and revisions to this model that should happen, but they're for another branch.