example notebooks
Notebooks featuring the bequest portfolio consumer type almost complete.

There is a minor issue in the notebooks: for some reason, the accidental and terminal bequest portfolio types are returning the same consumption functions and wealth moments. This obviously should not occur, since the dictionary creates instances with different parameter values.