Add "refinement loop" to basic portfolio solver
The proper human wealth and lower bound of MPC calculations are now in the "advanced" portfolio solver in ConsPortfolioModel, but not actually used in the consumption function. Calculation of the limiting share has been tweaked to be more accurate.

The "basic" solver in ConsRiskyAssetModel now goes through a loop to refine its search for optimal portfolio share at each aNrm gridpoint by "zooming in". We will probably want to add a parameter to specify how many times to do this loop, but it's currently set to 3.