Fix Markov tests and examples
Adjusted base dictionary for one test, changed parameter formatting for the other. In the example notebook, custom built MrkvArray has to be passed to the AgentType instance *after* initialization, because MrkvArray now has a default constructor. This behavior can be improved on later once there's just one call to construct() (and improved dictionaries).