ConsIndShock solver draws on infimum of IncShkDstn
I made these changes almost three weeks ago but forgot to commit. The solver for ConsIndShock now uses information in the limit dictionary of IncShkDstn when determining the "worst income shock", under the assumption that the "joint infimum" exists. As a default backup, it uses the lowest discretized shock, like before.

This does not really change the behavior of the solution unless you both add points *extremely* close to zero into aXtraGrid *and* actually include the infimum (as a tiny point mass) in the discretization itself.