Remove update_Rfree()
For a while now, there was a bit of code in PerfForesight that checked whether Rfree was a singleton or a list, and then set things up for that. This created a lot of work for something that had a very small solution: just put brackets around Rfree in parameter dictionaries.

This commit removes that method and fixes all of the dictionaries in the code. All tests pass, but I assume some notebooks will fail. There might also be some simulation code to simplify.