Added indexed Markov objects/events
Parameters and distributions can now be indexed using square brackets. Markov matrix transitions can be designated with j ~ {ArrayName}(i), where i is the old state and j is the new state.

Functions can't be indexed like this, but I added a small new interpolator that remaps cFunc[i](x) to cFunc(i,x) etc. In ConsMarkovModel, it's listed as cFuncX in the solution.