Add alternate notation for metadata flags
Objects can now be marked as a pre-state, part of the solution, offset in time, or a common/shared variable by adding a special symbol in the declaration. The old way of naming them explicitly in the appropriate entry also still works. The "common" designation is collected but not implemented yet.