Random variables can be marked as common
Variables declared as common or marked with an & in their declaration will be treated as non-idiosyncratic. If they are in the assignment of any kind of random event, a single common value will be assigned to all agents.

BUG that I just realized: This will not work properly with lifecycle models, because different random number generators are used for each age. Need to think about this.