First pass on transition matrix constructors
Wrote most of the code to build transition matrices on discretized state spaces. Committing now because I'm about to implement changes to a method that might break it.

As is, arrival-to-arrival transition matrices are stored in the trans_arrays attribute of the simulator. These go from the arrival state space in period t to the arrival state space in period t+1, including the replacement of dead agents in case of a loop.

Arrays for translating the distribution of arrival states into the distribution of *any other requested variable* are stored in the matrices attribute of each period.

All of these names will likely be changed, I just put them there as a starting point.