Made basic SSJ constructor
Has only been tested with ConsIndShock, but (in theory) should work for other models as long as they have a model file. This is the "basic" version that only works on "one period infinite horizon" models. It also doesn't work properly when there's death and replacement, but that doesn't exist in the original SSJ paper, so...

Still to do: improve grid specifications / options, test with other models, try to speed up the code a bit. Also test against current SSJ code.