Update sim_birth for ConsAggShockModel
Also had to change two target values because of the change in initial conditions. Testing showed that this discrepancy is entirely due to RNG changes. Copy-pasting the old sim_birth code into the method for AggShockConsumerType *almost* gets back to the prior test values. That is, just changing *how* the RNG is invoked perturbs the "solution" of the model slightly.

Note that these tests use a very short timeframe and small population, so they are extra sensitive to initial conditions. Hence "solution" in quotes.