Simplify setting constructors at initialization
As is, setting a non-default constructor at instantiation is a little obnoxious, because the user has to deepcopy the constructor dictionary, change some entries, and then put the altered copy into their parameter dictionary.

This commit makes it much easier to set non-default constructor functions at initialization. The user *only* needs to put entries into params["constructors"] for the constructors they want to change; the others will use their default. If the user wants to turn off a constructor with a default, they can set that entry to None.