Legacy simulator works for ConsExtMargMed
Simulation seemingly works as intended using the old/legacy simulator. There is one known bug/hitch: pLvlNextFunc is "offset" in the parent model but not in this model. New simulator system will correctly handle it, but I'll probably need to add a new method just to deal with this tiny difference.