Move notebooks out of /examples/ConsumptionSaving/
The directory structure for model documentation was inconsistent: some modules had their own subdirectory with a notebook for each AgentType subclass, others were shoved in /examples/ConsumptionSaving/, and others had an old/outdated notebook in the latter along with a newer one in the former.

This commit deletes old notebooks from /examples/ConsumptionSaving/ and moves the remainder into appropriately named subdirectories (and changes their names). A bunch of them need substantial work, but that's not for this PR.