Collapse fixed share model into RiskyAsset
We had two models that were nearly identical: one in which the agents are stuck at RiskyShare=1.0, and another in which they were stuck at some RiskyShare, which may or may not be 1.0. The "fixed share" solver had its name changed to replace the original RiskyAsset solver, which has been deleted. One parameter was added to the AgentType to make it work.

Tests pass for the model file, but need to see if all examples work.